requirements.txt
aiohttp==3.8.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord==2.0.0
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
load-dotenv==0.1.0
multidict==5.2.0
python-dotenv==0.21.0
yarl==1.8.1
